Refactor physical backup controller for better extensibility
